rpc: call CreateNewBlock via miner interface
This makes the options argument mandatory for BlockAssembler constructor,
dropping implicit handling of ArgsManager. The caller i.e. the Mining
interface implementation now handles this.

In Stratum v2 the pool communicates how many extra bytes it needs for
its own outputs (payouts, extra commitments, etc). This needs to be
substracted from what the user set as -blockmaxweight.

To achieve that the caller would have to pass in an options object,
and not forget to also process -blockmintxfee.